front sin secrets en settings.py

+++ Equivale/productos/models.py
@@ -85,7 +85,7 @@ class Pedido(models.Model):
 class TipoArbol(models.Model):
     nombre = models.CharField(max_length=100)
     descripcion = models.TextField(blank=True, null=True)
-    costo = models.DecimalField(max_digits=10, decimal_places=2)
+    costo = models.IntegerField()
     def __str__(self):
         return self.nombre
